Improve packages location
