+++ eternalcore/src/main/java/com/eternalcode/core/EternalCore.java
@@ -41,19 +41,19 @@ import com.eternalcode.core.command.contextual.ViewerContextual;
 import com.eternalcode.core.command.handler.InvalidUsage;
 import com.eternalcode.core.command.handler.NotificationHandler;
 import com.eternalcode.core.command.handler.PermissionMessage;
-import com.eternalcode.core.command.implementation.AlertCommand;
-import com.eternalcode.core.command.implementation.EnchantCommand;
-import com.eternalcode.core.command.implementation.FeedCommand;
-import com.eternalcode.core.command.implementation.FlyCommand;
-import com.eternalcode.core.command.implementation.GameModeCommand;
-import com.eternalcode.core.command.implementation.GiveCommand;
-import com.eternalcode.core.command.implementation.GodCommand;
-import com.eternalcode.core.command.implementation.HatCommand;
-import com.eternalcode.core.command.implementation.HealCommand;
-import com.eternalcode.core.command.implementation.KillCommand;
-import com.eternalcode.core.command.implementation.RepairCommand;
-import com.eternalcode.core.command.implementation.SkullCommand;
-import com.eternalcode.core.command.implementation.SpeedCommand;
+import com.eternalcode.core.command.implementation.info.AlertCommand;
+import com.eternalcode.core.command.implementation.item.EnchantCommand;
+import com.eternalcode.core.command.implementation.player.FeedCommand;
+import com.eternalcode.core.command.implementation.player.FlyCommand;
+import com.eternalcode.core.command.implementation.player.GameModeCommand;
+import com.eternalcode.core.command.implementation.item.GiveCommand;
+import com.eternalcode.core.command.implementation.player.GodCommand;
+import com.eternalcode.core.command.implementation.item.HatCommand;
+import com.eternalcode.core.command.implementation.player.HealCommand;
+import com.eternalcode.core.command.implementation.player.KillCommand;
+import com.eternalcode.core.command.implementation.item.RepairCommand;
+import com.eternalcode.core.command.implementation.item.SkullCommand;
+import com.eternalcode.core.command.implementation.player.SpeedCommand;
 import com.eternalcode.core.command.implementation.info.OnlinePlayerCountCommand;
 import com.eternalcode.core.command.implementation.info.OnlinePlayersListCommand;
 import com.eternalcode.core.command.implementation.info.PingCommand;
@@ -82,7 +82,7 @@ import com.eternalcode.core.configuration.ConfigurationManager;
 import com.eternalcode.core.configuration.implementation.LocationsConfiguration;
 import com.eternalcode.core.configuration.implementation.PlaceholdersConfiguration;
 import com.eternalcode.core.configuration.implementation.PluginConfiguration;
-import com.eternalcode.core.configuration.language.LanguageConfiguration;
+import com.eternalcode.core.language.LanguageConfiguration;
 import com.eternalcode.core.database.DatabaseManager;
 import com.eternalcode.core.database.NoneRepository;
 import com.eternalcode.core.database.wrapper.HomeRepositoryOrmLite;
