+++ eternalcore/src/main/java/com/eternalcode/core/command/implementation/item/SkullCommand.java
@@ -1,4 +1,4 @@
-package com.eternalcode.core.command.implementation;
+package com.eternalcode.core.command.implementation.item;
 
 import com.eternalcode.core.chat.notification.NoticeService;
 import dev.rollczi.litecommands.argument.Arg;
