+++ eternalcore/src/main/java/com/eternalcode/core/command/implementation/player/FeedCommand.java
@@ -1,4 +1,4 @@
-package com.eternalcode.core.command.implementation;
+package com.eternalcode.core.command.implementation.player;
 
 import com.eternalcode.core.viewer.Viewer;
 import com.eternalcode.core.chat.notification.NoticeService;
@@ -11,23 +11,22 @@ import dev.rollczi.litecommands.command.permission.Permission;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-@Route(name = "fly")
-@Permission("eternalcore.fly")
-public class FlyCommand {
+@Route(name = "feed")
+@Permission("eternalcore.feed")
+public class FeedCommand {
 
     private final NoticeService noticeService;
 
-    public FlyCommand(NoticeService noticeService) {
+    public FeedCommand(NoticeService noticeService) {
         this.noticeService = noticeService;
     }
 
     @Execute
     void execute(Viewer audience, CommandSender sender, @Arg @By("or_sender") Player player) {
-        player.setAllowFlight(!player.getAllowFlight());
+        player.setFoodLevel(20);
 
         this.noticeService.create()
-            .notice(messages -> messages.other().flyMessage())
-            .placeholder("{STATE}", messages -> player.getAllowFlight() ? messages.format().enable() : messages.format().disable())
+            .notice(messages -> messages.other().feedMessage())
             .player(player.getUniqueId())
             .send();
 
@@ -36,10 +35,10 @@ public class FlyCommand {
         }
 
         this.noticeService.create()
-            .notice(messages -> messages.other().flySetMessage())
             .placeholder("{PLAYER}", player.getName())
-            .placeholder("{STATE}", messages -> player.getAllowFlight() ? messages.format().enable() : messages.format().disable())
+            .notice(messages -> messages.other().feedMessageBy())
             .viewer(audience)
             .send();
     }
 }
+
