+++ eternalcore/src/main/java/com/eternalcode/core/configuration/ConfigurationManager.java
@@ -3,7 +3,7 @@ package com.eternalcode.core.configuration;
 import com.eternalcode.core.chat.notification.Notification;
 import com.eternalcode.core.configuration.composer.DurationComposer;
 import com.eternalcode.core.configuration.composer.NotificationComposer;
-import com.eternalcode.core.configuration.language.LanguageComposer;
+import com.eternalcode.core.language.LanguageComposer;
 import com.eternalcode.core.configuration.composer.PositionComposer;
 import com.eternalcode.core.language.Language;
 import com.eternalcode.core.shared.Position;
