+++ eternalcore/src/main/java/com/eternalcode/core/language/LanguageComposer.java
@@ -1,4 +1,4 @@
-package com.eternalcode.core.configuration.language;
+package com.eternalcode.core.language;
 
 import com.eternalcode.core.configuration.composer.SimpleComposer;
 import com.eternalcode.core.language.Language;
