+++ eternalcore/src/main/java/com/eternalcode/core/language/LanguageSelector.java
@@ -1,6 +1,5 @@
-package com.eternalcode.core.configuration.language;
+package com.eternalcode.core.language;
 
-import com.eternalcode.core.language.Language;
 import com.google.common.collect.ImmutableList;
 import net.dzikoysk.cdn.entity.Contextual;
 import net.dzikoysk.cdn.entity.Description;
