+++ eternalcore/src/main/java/com/eternalcode/core/language/LocaleComposer.java
@@ -1,4 +1,4 @@
-package com.eternalcode.core.configuration.language;
+package com.eternalcode.core.language;
 
 import com.eternalcode.core.configuration.composer.SimpleComposer;
 import panda.std.Result;
